# examples/books/1.0.0/books/target/generated-sources/openapi/src/main/java/dtu/services/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.19.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package dtu.services.api;

import java.util.List;

import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import dtu.services.api.model.Bog;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-27T10:25:24.903746840Z[Etc/UTC]", comments = "Generator version: 7.19.0")
@Validated
@Tag(name = "books", description = "the books API")
public interface BooksApi {

    String PATH_ADD = "/books";
    /**
     * POST /books : Register a new book
     * Adds a new book record to the library. The request body must contain the valid book data model. Use this to expand the inventory. 
     *
     * @param bog The book object that needs to be persisted. (required)
     * @return Book successfully created and saved. (status code 201)
     */
    @Operation(
        operationId = "add",
        summary = "Register a new book",
        description = "Adds a new book record to the library. The request body must contain the valid book data model. Use this to expand the inventory. ",
        responses = {
            @ApiResponse(responseCode = "201", description = "Book successfully created and saved.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Bog.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_ADD,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Bog> add(
        @Parameter(name = "Bog", description = "The book object that needs to be persisted.", required = true) @Valid @RequestBody Bog bog
    );


    String PATH_GET_ALL = "/books";
    /**
     * GET /books : List all available books
     * Returns a complete list of all books currently registered in the inventory. Useful for browsing the collection or performing a broad search. 
     *
     * @return A list of book objects. (status code 200)
     */
    @Operation(
        operationId = "getAll",
        summary = "List all available books",
        description = "Returns a complete list of all books currently registered in the inventory. Useful for browsing the collection or performing a broad search. ",
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of book objects.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Bog.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_ALL,
        produces = { "application/json" }
    )
    ResponseEntity<List<Bog>> getAll(
        
    );


    String PATH_GET_BY_ID = "/books/{id}";
    /**
     * GET /books/{id} : Fetch specific book details
     * Retrieves comprehensive information about a single book using its unique 64-bit integer identifier. Use this when you have a specific ID and need the full metadata (title, author, etc.). 
     *
     * @param id The unique numeric ID of the book (e.g., 101). (required)
     * @return Successfully retrieved the book details. (status code 200)
     *         or No book was found with the provided ID. (status code 404)
     */
    @Operation(
        operationId = "getById",
        summary = "Fetch specific book details",
        description = "Retrieves comprehensive information about a single book using its unique 64-bit integer identifier. Use this when you have a specific ID and need the full metadata (title, author, etc.). ",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved the book details.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Bog.class))
            }),
            @ApiResponse(responseCode = "404", description = "No book was found with the provided ID.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<Bog> getById(
        @NotNull @Parameter(name = "id", description = "The unique numeric ID of the book (e.g., 101).", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );

}